+++ src/Flare.Theme.FluentUI2.Tokens/FluentUI2Tokens.cs
@@ -1290,6 +1290,14 @@ public class FluentUI2Tokens
     internal static readonly TimePickerTokens TimePicker = new()
     {
         ColumnsSepSize = "1.5rem",
+        DialCenterSize = "0.5rem",
+        DialHandleSize = "2.25rem",
+        DialSize = "16rem",
+        DialTrackWidth = "0.125rem",
+        PeriodHeight = "auto",
+        PeriodWidth = "auto",
+        TimeFieldHeight = "auto",
+        TimeFieldWidth = "5rem",
         DisplaySize = "2.75rem",
         HeadlineTracking = "0.05em",
         PanelRadius = "var(--flare-shape-extra-large)",